requirements.txt
google-genai>=1.0.0
langchain-google-genai>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.20.0
supabase>=2.0.0
apscheduler>=3.10.0
